+++ src/Pwa/PageResult/Navigation/NavigationPageResult.php
@@ -24,11 +24,6 @@ class NavigationPageResult extends AbstractPageResult
     protected $breadcrumb;
 
     /**
-     * @var array
-     */
-    protected $listingConfiguration;
-
-    /**
      * @return CmsPageEntity|null
      */
     public function getCmsPage(): ?CmsPageEntity
@@ -52,19 +47,6 @@ class NavigationPageResult extends AbstractPageResult
     public function setBreadcrumb(array $breadcrumb)
     {
         $this->breadcrumb = $breadcrumb;
-    }
-
-    /**
-     * @return array
-     */
-    public function getListingConfiguration(): array
-    {
-        return $this->listingConfiguration;
-    }
-
-    public function setListingConfiguration(array $listingConfiguration)
-    {
-        $this->listingConfiguration = $listingConfiguration;
     }
 
     /**