+++ src/Pwa/PageResult/Navigation/NavigationPageResultHydrator.php
@@ -6,11 +6,19 @@ use Shopware\Core\Content\Category\CategoryEntity;
 use Shopware\Core\Content\Cms\CmsPageEntity;
 use Shopware\Core\Content\Product\SalesChannel\Listing\ProductListingResult;
 use Shopware\Core\Content\Property\PropertyGroupEntity;
+use Shopware\Core\Content\Seo\SeoUrl\SeoUrlEntity;
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\AggregationResult\AggregationResultCollection;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\AggregationResult\Metric\EntityResult;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
 use Shopware\Storefront\Framework\Routing\Router;
 use SwagShopwarePwa\Pwa\Controller\PageController;
 use SwagShopwarePwa\Pwa\PageLoader\Context\PageLoaderContext;
+use SwagShopwarePwa\Pwa\PageLoader\NavigationPageLoader;
 use SwagShopwarePwa\Pwa\PageResult\Navigation\AggregationResultHydrator\AggregationResultHydratorInterface;
 
 class NavigationPageResultHydrator
@@ -21,26 +29,21 @@ class NavigationPageResultHydrator
     private $pageResult;
 
     /**
+     * @var EntityRepositoryInterface
+     */
+    private $seoUrlRepository;
+
+    /**
      * @var Router
      */
     private $router;
 
-    /**
-     * @var AggregationResultHydratorInterface[]
-     */
-    private $aggregationResultHydrators;
-
-    public function __construct(Router $router, iterable $aggregationResultHydrators)
+    public function __construct(Router $router, EntityRepositoryInterface $seoUrlRepository)
     {
         $this->router = $router;
+        $this->seoUrlRepository = $seoUrlRepository;
 
         $this->pageResult = new NavigationPageResult();
-
-        /** @var AggregationResultHydratorInterface[] $aggregationResultHydrators */
-        foreach($aggregationResultHydrators as $resultHydrator)
-        {
-            $this->aggregationResultHydrators[$resultHydrator->getSupportedAggregationType()] = $resultHydrator;
-        }
     }
 
     public function hydrate(PageLoaderContext $pageLoaderContext, CategoryEntity $category, ?CmsPageEntity $cmsPageEntity): NavigationPageResult
@@ -50,98 +53,55 @@ class NavigationPageResultHydrator
 
         $this->setBreadcrumbs(
             $category,
-            $pageLoaderContext->getContext()->getSalesChannel()->getNavigationCategoryId()
+            $pageLoaderContext->getContext()
         );
 
         $this->pageResult->setResourceType($pageLoaderContext->getResourceType());
         $this->pageResult->setResourceIdentifier($pageLoaderContext->getResourceIdentifier());
 
-        $this->pageResult->setListingConfiguration($this->getAvailableFilters());
-
         return $this->pageResult;
     }
 
-    private function getAvailableFilters(): array
-    {
-        if($this->pageResult->getCmsPage() === null)
-        {
-            return [];
-        }
-
-        // Assuming a page only has one listing
-        $listingSlot = $this->pageResult->getCmsPage()->getFirstElementOfType('product-listing');
-
-        if($listingSlot === null)
-        {
-            return [];
-        }
-
-        /** @var ProductListingResult $listing */
-        $listing = $listingSlot->getData()->getListing();
-        $filters = [];
-
-        $this->preparePropertyAggregations($listing->getAggregations());
-
-        foreach($listing->getAggregations() as $key => $aggregation)
-        {
-            $filters[$key] = $this->aggregationResultHydrators[get_class($aggregation)]->hydrate($aggregation);
-        }
-
-        $currentFilters = $listing->getCurrentFilters();
-
-        $listingConfig = [
-            'availableSortings' => $listing->getSortings(),
-            'availableFilters' => $filters,
-            'activeFilters' => $currentFilters,
-        ];
-
-        return $listingConfig;
-    }
-
-    private function setBreadcrumbs(CategoryEntity $category, string $rootCategoryId): void
+    private function setBreadcrumbs(CategoryEntity $category, SalesChannelContext $context): void
     {
         $breadcrumbs = [];
 
+        $rootCategoryId = $context->getSalesChannel()->getNavigationCategoryId();
+
         $categoryBreadcrumbs = $category->buildSeoBreadcrumb($rootCategoryId) ?? [];
 
-        foreach($categoryBreadcrumbs as $id => $name)
-        {
+        $canonicalUrls = $this->getCanonicalUrls(array_keys($categoryBreadcrumbs), $context);
+
+        foreach ($categoryBreadcrumbs as $id => $name) {
             $breadcrumbs[$id] = [
                 'name' => $name,
-                'path' => $this->router->generate(PageController::NAVIGATION_PAGE_ROUTE, ['navigationId' => $id]),
+                'path' => $canonicalUrls[$id] ?? $this->router->generate(PageController::NAVIGATION_PAGE_ROUTE, ['navigationId' => $id]),
             ];
         }
 
         $this->pageResult->setBreadcrumb($breadcrumbs);
     }
 
-    private function preparePropertyAggregations(AggregationResultCollection $aggregations): AggregationResultCollection
+    private function getCanonicalUrls(array $categoryIds, SalesChannelContext $context): array
     {
-        foreach($aggregations as $aggKey => $aggregation)
-        {
-            if($aggKey !== 'properties')
-            {
-                continue;
-            }
+        $criteria = new Criteria();
 
-            // For categories/listings without products
-            if(!$aggregation instanceof EntityResult)
-            {
-                continue;
-            }
+        $criteria->addFilter(new EqualsFilter('routeName', PageController::NAVIGATION_PAGE_ROUTE));
+        $criteria->addFilter(new EqualsFilter('isCanonical', true));
+        $criteria->addFilter(new EqualsAnyFilter('foreignKey', $categoryIds));
+        $criteria->addFilter(new EqualsFilter('salesChannelId', $context->getSalesChannelId()));
+        $criteria->addFilter(new EqualsFilter('languageId', $context->getContext()->getLanguageId()));
 
-            /** @var $aggregation EntityResult */
-            foreach($aggregation->getEntities() as $key => $propertyGroup)
-            {
-                /** @var PropertyGroupEntity $propertyGroup */
-                $result = new EntityResult($propertyGroup->getName(), $propertyGroup->getOptions());
-                $aggregations->add($result);
-            }
+        $result = $this->seoUrlRepository->search($criteria, $context->getContext());
 
-            $aggregations->remove($aggKey);
+        $pathByCategoryId = [];
 
+        /** @var SeoUrlEntity $seoUrl */
+        foreach($result as $seoUrl) {
+            // Map all urls to their corresponding category
+            $pathByCategoryId[$seoUrl->getForeignKey()] = '/' . ( $seoUrl->getSeoPathInfo() ?? $seoUrl->getPathInfo() );
         }
 
-        return $aggregations;
+        return $pathByCategoryId;
     }
 }