+++ src/Resources/app/administration/src/main.js
@@ -1,0 +1,1 @@
+import './module/sw-cms/component/sw-cms-sidebar/index';